fix: keep the mount tokens at the include root and update h2 for RUSTSEC-2026-0258
The mount tokens are printed inside the trait bounds a user reads when a slot is
left unbound, so splitting them into a submodule lengthened those diagnostics for
no gain; they move back beside the spine and the snapshots shrink to the file
path that genuinely changed. h2 goes to 0.4.16 for the unbounded empty DATA
frame advisory.
