+++ tests/ui/subscriber_out_capability_mismatch.stderr
@@ -34,7 +34,7 @@ note: required for `__RsOutDef_forward<SlotPublisher<MemoryPublisher, DefaultSlo
    |                                                     ------------ unsatisfied trait bound introduced here
    = note: required for `(ruststream::runtime::WithSource<MemoryPublish>,)` to implement `ruststream::runtime::SlotCommit<runtime::app::include::InjectMount, MemoryBroker, Identity, (), (), PublishIdentity, forward>`
 note: required by a bound in `IncludeSlots::<'_, Mount, B, Layers, C, State, Pipeline, Def, (ruststream::runtime::MissingSlot<M>,)>::publisher`
-  --> src/runtime/app/include.rs
+  --> src/runtime/app/include/slot_builder.rs
    |
    |     pub fn publisher<NewSource>(self, source: NewSource)
    |            --------- required by a bound in this associated function
