+++ tests/ui/subscriber_out_unbound_slot.stderr
@@ -17,7 +17,7 @@ error[E0277]: not every Out slot of this handler is bound
              `(ruststream::runtime::WithSource<A0>,)` implements `ruststream::runtime::SlotCommit<runtime::app::include::BatchInjectMount, B, Layers, C, State, Pipeline, Def>`
            and $N others
 note: required by a bound in `IncludeSlots::<'s, Mount, B, Layers, C, State, Pipeline, Def, Slots>::mount`
-  --> src/runtime/app/include.rs
+  --> src/runtime/app/include/slot_builder.rs
    |
    |     pub fn mount(self)
    |            ----- required by a bound in this associated function
